quantity selecting

+++ 2020/BillSplitter/BillSplitter/Data/OrderRepository.cs
@@ -20,7 +20,8 @@ namespace BillSplitter.Data
         public void DeleteByUserAndPosition(int userId, int positionId)
         {
             var order = _context.Orders.FirstOrDefault(o => (o.Customer.UserId==userId && o.PositionId==positionId ));
-            _context.Orders.Remove(order);
+            if(order!=null)
+                _context.Orders.Remove(order);
         }
     }
 }
